a few more fixes
